# perf: calltree speed up getalltypes