--- a/log-viewer/modules/components/calltree-view/CalltreeView.ts
+++ b/log-viewer/modules/components/calltree-view/CalltreeView.ts
@@ -181,7 +181,7 @@ export class CalltreeView extends LitElement {
                   <vscode-dropdown @change="${this._handleTypeFilter}">
                     <vscode-option>None</vscode-option>
                     ${repeat(
-                      this._getAllTypes(this.timelineRoot?.children ?? []),
+                      this._getAllTypes(this.timelineRoot?.children ?? []).sort(),
                       (type, _index) => html`<vscode-option>${type}</vscode-option>`,
                     )}
                   </vscode-dropdown>
@@ -200,10 +200,32 @@ export class CalltreeView extends LitElement {
   }
 
   _getAllTypes(data: LogLine[]): string[] {
-    const flatten = (line: LogLine): LogLine[] => [line, ...line.children.flatMap(flatten)];
-    const flattened = data.flatMap(flatten);
+    let len = data.length;
+    if (len === 0) {
+      return [];
+    }
 
-    return [...new Set(flattened.map((item) => item.type?.toString() ?? ''))].sort();
+    const types = new Set<string>();
+    const children: LogLine[] = [];
+
+    while (len--) {
+      const line = data[len];
+      if (!line) {
+        continue;
+      }
+      types.add(line.type?.toString() ?? '');
+      for (const c of line?.children ?? []) {
+        children.push(c);
+      }
+    }
+
+    if (children.length > 0) {
+      for (const c of this._getAllTypes(children)) {
+        types.add(c);
+      }
+    }
+
+    return Array.from(types);
   }
 
   _handleShowDetailsChange(event: Event) {